feat(desktop): move to general page
